this works too and is smaller, not finished. will be done as the time goes by. not critical
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@9792 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/write_web.c
@@ -2,7 +2,7 @@
 typedef struct
 {
   char *path;			/* Web page URL path */
-  unsigned long offset;		/* Web page data */
+  unsigned int offset;		/* Web page data */
   int size;			/* Size of web page in bytes */
 //  int csize;                  /* Size of web page in bytes */
 } websRomPageIndexType;
